feat: Export command
