+++ src/cli/notes.rs
@@ -1,9 +1,4 @@
-use std::{
-    collections::{HashMap, HashSet},
-    fs::File,
-    io::Write,
-    path::PathBuf,
-};
+use std::collections::{HashMap, HashSet};
 
 use clap::ValueEnum;
 use comfy_table::{presets, Attribute, Cell, ContentArrangement, Table};
@@ -18,7 +13,6 @@ use miden_objects::{
     notes::{NoteInputs, NoteMetadata},
     Digest,
 };
-use miden_tx::utils::Serializable;
 
 use super::{Client, Parser};
 use crate::cli::{
@@ -63,19 +57,6 @@ pub enum Notes {
         inputs: bool,
     },
 
-    /// Export note data to a binary file.
-    #[clap(short_flag = 'e')]
-    Export {
-        /// Note ID of the note to show. We only allow to export a note that has been created using
-        /// this client
-        #[clap()]
-        id: String,
-
-        /// Path to the file that will contain the note data. If not provided, the filename will be the input note ID
-        #[clap()]
-        filename: Option<PathBuf>,
-    },
-
     /// List consumable notes
     #[clap(short_flag = 'c')]
     ListConsumable {
@@ -109,10 +90,6 @@ impl Notes {
             },
             Notes::Show { id, script, vault, inputs } => {
                 show_note(client, id.to_owned(), *script, *vault, *inputs)?;
-            },
-            Notes::Export { id, filename } => {
-                export_note(&client, id, filename.clone())?;
-                println!("Succesfully exported note {}", id);
             },
             Notes::ListConsumable { account_id } => {
                 list_consumable_notes(client, account_id)?;
@@ -151,39 +128,6 @@ fn list_notes<N: NodeRpcClient, R: FeltRng, S: Store>(
 
     print_notes_summary(zipped_notes)?;
     Ok(())
-}
-
-// EXPORT NOTE
-// ================================================================================================
-pub fn export_note<N: NodeRpcClient, R: FeltRng, S: Store>(
-    client: &Client<N, R, S>,
-    note_id: &str,
-    filename: Option<PathBuf>,
-) -> Result<File, String> {
-    let note_id = Digest::try_from(note_id)
-        .map_err(|err| format!("Failed to parse input note id: {}", err))?
-        .into();
-    let output_note = client
-        .get_output_notes(miden_client::store::NoteFilter::Unique(note_id))?
-        .pop()
-        .expect("should have an output note");
-
-    // Convert output note into input note before exporting
-    let input_note: InputNoteRecord = output_note
-        .try_into()
-        .map_err(|_err| format!("Can't export note with ID {}", note_id.to_hex()))?;
-
-    let file_path = filename.unwrap_or_else(|| {
-        let mut dir = PathBuf::new();
-        dir.push(note_id.inner().to_string());
-        dir
-    });
-
-    let mut file = File::create(file_path).map_err(|err| err.to_string())?;
-
-    file.write_all(&input_note.to_bytes()).map_err(|err| err.to_string())?;
-
-    Ok(file)
 }
 
 // SHOW NOTE
@@ -462,7 +406,7 @@ mod tests {
     };
     use uuid::Uuid;
 
-    use crate::cli::{get_input_note_with_id_prefix, import::import_note, notes::export_note};
+    use crate::cli::{export::export_note, get_input_note_with_id_prefix, import::import_note};
 
     #[tokio::test]
     async fn test_import_note_validation() {
